# Reduce allocations in StringBuilder.AppendFormat for primitive types