--- a/src/mscorlib/shared/System/ISpanFormattable.cs
+++ b/src/mscorlib/shared/System/ISpanFormattable.cs
@@ -1,0 +1,11 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace System
+{
+    internal interface ISpanFormattable
+    {
+        bool TryFormat(Span<char> destination, out int charsWritten, string format = null, IFormatProvider provider = null);
+    }
+}