Use the new `TryFormat` APIs to avoid string allocations for primitive types inside `StringBuilder.AppendFormat`, used by `string.Format`/interpolated strings.

Benchmark:

```c#
using System;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Attributes.Jobs;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
[InProcess]
public class Program
{
    public static void Main() => BenchmarkRunner.Run<Program>();

    public static readonly Guid s_guid = new Guid("8759798E-3430-47AA-9A53-C6813F1A7456");

    [Benchmark]
    public void StringFormat()
    {
        Guid guid = s_guid;
        int value = 42;
        for (int i = 0; i < 100_000; i++) string.Format("{0} blah {1}", guid, value);
    }
}
```

Before:

```
       Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
------------- |---------:|----------:|----------:|----------:|----------:|
 StringFormat | 20.78 ms | 0.1334 ms | 0.1183 ms | 7437.5000 |  29.75 MB |
```

After:

```
       Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
------------- |---------:|----------:|----------:|----------:|----------:|
 StringFormat | 19.80 ms | 0.0735 ms | 0.0688 ms | 4187.5000 |  16.78 MB |
```

Additional improvements can be looked into subsequently, such as avoiding any `itemFormat` string allocations.