@stephentoub 2017-11-20
@dotnet-bot test CentOS7.1 x64 Checked Innerloop Build and Test please ("Could not load type 'System.MemoryExtensions'")