@stephentoub 2017-11-21
@jkotas, it's ok to merge with these known type load exceptions on centos, right?